utils: move linspace utils to avoid circular imports
