+++ orangecontrib/spectroscopy/widgets/owhyper.py
@@ -26,6 +26,7 @@ from Orange.widgets.utils import saveplot
 from Orange.data import DiscreteVariable, ContinuousVariable
 
 from orangecontrib.spectroscopy.preprocess import Integrate
+from orangecontrib.spectroscopy.utils import values_to_linspace, index_values
 
 from orangecontrib.spectroscopy.widgets.owspectra import InteractiveViewBox, \
     MenuFocus, CurvePlot, SELECTONE, SELECTMANY, INDIVIDUAL, AVERAGE, \
@@ -94,42 +95,6 @@ def refresh_integral_markings(dis, markings_list, curveplot):
                 dot.setPen(pg.mkPen(color=QColor(color), width=5))
                 dot.setZValue(10)
                 add_marking(dot)
-
-
-def values_to_linspace(vals):
-    """Find a near maching linspace for the values given.
-    The problem is that some values can be missing and
-    that they are inexact. The minumum and maximum values
-    are kept as limits."""
-    vals = vals[~np.isnan(vals)]
-    if len(vals):
-        vals = np.unique(vals)  # returns sorted array
-        if len(vals) == 1:
-            return vals[0], vals[0], 1
-        minabsdiff = (vals[-1] - vals[0])/(len(vals)*100)
-        diffs = np.diff(vals)
-        diffs = diffs[diffs > minabsdiff]
-        first_valid = diffs[0]
-        # allow for a percent mismatch
-        diffs = diffs[diffs < first_valid*1.01]
-        step = np.mean(diffs)
-        size = int(round((vals[-1]-vals[0])/step) + 1)
-        return vals[0], vals[-1], size
-    return None
-
-
-def location_values(vals, linspace):
-    vals = np.asarray(vals)
-    if linspace[2] == 1:  # everything is the same value
-        width = 1
-    else:
-        width = (linspace[1] - linspace[0]) / (linspace[2] - 1)
-    return (vals - linspace[0]) / width
-
-
-def index_values(vals, linspace):
-    """ Remap values into index of array defined by linspace. """
-    return np.round(location_values(vals, linspace)).astype(int)
 
 
 def _shift(ls):
